+++ src/components/map/map.constants.ts
@@ -1,3 +1,5 @@
 export const URL_MARKER_DEFAULT = '../img/pin.svg';
 
 export const URL_MARKER_CURRENT = '../img/pin-active.svg';
+
+export const SIZE_MARKER = [27, 39];
